refactor(types): Rename `overlay.ts` to `alignment.ts` for clarity
