+++ src/types/index.ts
@@ -1,5 +1,5 @@
+export * from './alignment';
 export * from './common';
 export * from './html-tag';
-export * from './overlay';
 export * from './polymorphic';
 export * from './styling';
